+++ CHANGES.md
@@ -1,5 +1,6 @@
 #### [unreleased]
 * ensure `$slug` is set in `Init->rename_on_activation()`
+* specify branch for changelog
 
 #### 8.4.2 / 2018-11-01
 * updated password fields to not autoload saved passwords, thanks @figureone