+++ CHANGES.md
@@ -6,6 +6,7 @@
 * refactored GitHub release asset code to get direct download link
 * refactored Bitbucket release asset code to get redirected download link for AWS
 * refactored GitLab release asset code to get redirected download link
+* exit early if checking _View details_ but not done with background update, avoids PHP notices
 
 #### 8.4.2 / 2018-11-01
 * updated password fields to not autoload saved passwords, thanks @figureone